# api/swagger.yaml
openapi: 3.0.0
info:
  title: Aloux Platform API
  description: APIs para Plataforma Aloux
  version: 1.0.0
  contact:
    name: API Aloux
    url: https://aloux.mx
    email: developer@aloux.mx
servers:
  - url: http://localhost
    description: DEV
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      properties:
        name:
          type: string
          example: Arturo
        description:
          type: string
          example: Arturo Desc
        taxInfo:
          type: object
          properties:
            rfc:
              type: string
            name:
              type: string
            regime:
              type: object
            CFDIuse:
              type: object
            zip:
              type: string
            status:
              type: string
        files:
          type: object
      required:
        - name
    ######################################## Project  ########################################
...
paths:
...
  /client:
    post:
      summary: Crea un cliente.
      tags:
        - Client
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        "201": # status code
          description: A JSON array of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    get:
      summary: Obtiene los clientes.
      tags:
        - Client
      security:
        - bearerAuth: []
      responses:
        "200": # status code
          description: A JSON object of a clients
          content:
            application/json:
              schema:
                type: object
                items:
                  type: string
...
  /client/{CLIENT_ID}:
    get:
      parameters:
        - name: CLIENT_ID
          in: path
          description: id de un cliente
          required: true
          schema:
            type: string
            format: objectId
      tags:
        - Client
      summary: Obtiene el detalle de un cliente
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
        "400":
          description: bad request
    patch:
      parameters:
        - name: CLIENT_ID
          in: path
          description: id de un cliente
          required: true
          schema:
            type: string
      tags:
        - Client
      summary: Actualiza uno o varios campos de un cliente
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Client"
      responses:
        "202":
          description: accept

        "400":
          description: bad request
    delete:
      parameters:
        - name: CLIENT_ID
          in: path
          description: id del cliente
          required: true
          schema:
            type: string
      tags:
        - Client
      summary: Eliminar un cliente
      security:
        - bearerAuth: []
      responses:
        "200":
          description: ok
        "400":
          description: bad request
...